chore: Refactor workspace-operator, split comps
